Refactored and added a few packages
