+++ src/main/java/dk/dtu/compute/se/pisd/roborally/controller/boardElementController/BoardLaserController.java
@@ -1,11 +1,11 @@
 package dk.dtu.compute.se.pisd.roborally.controller.boardElementController;
 
 import dk.dtu.compute.se.pisd.roborally.interfaces.ILaser;
-import dk.dtu.compute.se.pisd.roborally.model.Command;
+import dk.dtu.compute.se.pisd.roborally.model.enums.Command;
 import dk.dtu.compute.se.pisd.roborally.model.CommandCard;
 import dk.dtu.compute.se.pisd.roborally.model.Player;
-import dk.dtu.compute.se.pisd.roborally.model.Space;
-import dk.dtu.compute.se.pisd.roborally.model.boardElement.activationElements.BoardLaser;
+import dk.dtu.compute.se.pisd.roborally.model.board.Space;
+import dk.dtu.compute.se.pisd.roborally.model.board.boardElement.activationElements.BoardLaser;
 
 public class BoardLaserController implements IBoardElementController, ILaser {
     private BoardLaser model;
