+++ src/main/java/dk/dtu/compute/se/pisd/roborally/interfaces/IActivateable.java
@@ -1,6 +1,6 @@
 package dk.dtu.compute.se.pisd.roborally.interfaces;
 
-import dk.dtu.compute.se.pisd.roborally.model.boardElement.activationElements.ActivationElement;
+import dk.dtu.compute.se.pisd.roborally.model.board.boardElement.activationElements.ActivationElement;
 
 /**
  * <p>Represents something that can be activated, e.g. an {@link ActivationElement} or a robot laser.</p>
