+++ src/main/java/dk/dtu/compute/se/pisd/roborally/interfaces/ILaser.java
@@ -1,7 +1,5 @@
 package dk.dtu.compute.se.pisd.roborally.interfaces;
 
-import dk.dtu.compute.se.pisd.roborally.model.Board;
-
 public interface ILaser extends IActivateable{
 
     /**
