+++ src/main/java/dk/dtu/compute/se/pisd/roborally/model/CommandCard.java
@@ -22,6 +22,7 @@
 package dk.dtu.compute.se.pisd.roborally.model;
 
 import dk.dtu.compute.se.pisd.designpatterns.observer.Subject;
+import dk.dtu.compute.se.pisd.roborally.model.enums.Command;
 import org.jetbrains.annotations.NotNull;
 
 /**
