+++ src/main/java/dk/dtu/compute/se/pisd/roborally/model/enums/Phase.java
@@ -19,7 +19,7 @@
  *  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  *
  */
-package dk.dtu.compute.se.pisd.roborally.model;
+package dk.dtu.compute.se.pisd.roborally.model.enums;
 
 /**
  * <p>Represents the current state of the program.</p>
