+++ src/main/java/dk/dtu/compute/se/pisd/roborally/view/BoardElementView.java
@@ -2,7 +2,7 @@ package dk.dtu.compute.se.pisd.roborally.view;
 
 import dk.dtu.compute.se.pisd.designpatterns.observer.Subject;
 import dk.dtu.compute.se.pisd.roborally.imageFinder.BoardElementImageFinder;
-import dk.dtu.compute.se.pisd.roborally.model.boardElement.BoardElement;
+import dk.dtu.compute.se.pisd.roborally.model.board.boardElement.BoardElement;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.BorderPane;
